+++ src/interface/metadata.hpp
@@ -113,7 +113,9 @@
   /** the variable participate in GMG calculations */                                    \
   PARTHENON_INTERNAL_FOR_FLAG(GMGProlongate)                                             \
   /** the variable participate in GMG calculations */                                    \
-  PARTHENON_INTERNAL_FOR_FLAG(GMGRestrict)
+  PARTHENON_INTERNAL_FOR_FLAG(GMGRestrict)                                               \
+  /** the variable must always be allocated for new blocks **/                           \
+  PARTHENON_INTERNAL_FOR_FLAG(ForceAllocOnNewBlocks)
 namespace parthenon {
 
 namespace internal {
@@ -234,7 +236,9 @@ class Metadata {
     FlagCollection(MetadataFlag first, Args... args)
         : FlagCollection({first, std::forward<Args>(args)...}, false) {}
     // Check if set empty
-    bool Empty() const { return (unions_.empty() && intersections_.empty()); }
+    bool Empty() const {
+      return (unions_.empty() && intersections_.empty() && exclusions_.empty());
+    }
     // Union
     template <template <class...> class Container_t, class... extra>
     void TakeUnion(const Container_t<MetadataFlag, extra...> &flags) {
@@ -335,6 +339,9 @@ class Metadata {
   static MetadataFlag AddUserFlag(const std::string &name);
   static bool FlagNameExists(const std::string &flagname);
   static MetadataFlag GetUserFlag(const std::string &flagname);
+  static MetadataFlag GetOrAddFlag(const std::string &name) {
+    return FlagNameExists(name) ? GetUserFlag(name) : AddUserFlag(name);
+  }
   static int num_flags;
 
   // Sparse threshold routines
@@ -543,13 +550,15 @@ class Metadata {
     PARTHENON_REQUIRE_THROWS(IsRefined(), "Variable must be registered for refinement");
     return refinement_funcs_;
   }
-  template <class ProlongationOp, class RestrictionOp>
+  template <class ProlongationOp, class RestrictionOp,
+            class InternalProlongationOp = refinement_ops::ProlongateInternalAverage>
   void RegisterRefinementOps() {
     PARTHENON_REQUIRE_THROWS(
         IsRefined(),
         "Variable must be registered for refinement to accept custom refinement ops");
     refinement_funcs_ =
-        refinement::RefinementFunctions_t::RegisterOps<ProlongationOp, RestrictionOp>();
+        refinement::RefinementFunctions_t::RegisterOps<ProlongationOp, RestrictionOp,
+                                                       InternalProlongationOp>();
   }
 
   // Operators